backintime; incorporate suggestions from k0Do
The configure script for the gnome GUI does not work for --no-check, meaning
gnome-session is always enforced. Not good for Xfce, Enlightenment etc.

Here my proposal for the PKGBUILd part:

GUI_ENFORCE=1 # set to 0 if you don't want a GUI at all
if [ -e '/usr/lib/libkonq.so' ]; then
  # KDE is installed
  depends=( ${depends[@]} xorg-utils python2-pyqt kdebindings-python )
  _type=( ${_type[@]} kde4 )
fi
if [ -n "`which gnome-session 2> /dev/null`" ]; then
  # GNOME really
  depends=( ${depends[@]} pygtk python-notify gnome-python meld gksu gnome-session )
  _type=( ${_type[@]} gnome )
fi
if [ ${GUI_ENFORCE} == 1 -a ${#_type[@]} == 0 ]; then
  # other DE like Xfce, Enlightenment etc.
  depends=( ${depends[@]} pygtk python-notify gnome-python meld gksu )
  _type=( ${_type[@]} gnome )
fi

Replace the line "./configure" with "cp Makefile.template Makefile" as a work
around for the gnome section. The check for gnome-session is useless here.
http://bazaar.launchpad.net/~bit-team/backintime/trunk/view/head:/gnome/configure